Modify PDB reading to rely on BioPython
Remove functions like parsing coordinates or missing residues. All this information can be accessed through biophython object.
Mol type check function is retained to prevent processing mixed complexes. This is only a preventive measure until the code is tested against such complexes, it will be mainted. However, with Biopython's polypeptide class we don't need to worry about it because whichever residue is left, it would belong to a protein.